cern_mag_llmops/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-core>=0.1.0
pypdf>=3.17.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
pytest>=7.4.0
chromadb>=0.4.22
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
tiktoken>=0.5.1 